+++ config.py
@@ -1,0 +1,11 @@
+import os
+
+API_KEYS = {
+    "OPENAI": os.getenv("OPENAI_API_KEY"),
+    "ANTHROPIC": os.getenv("ANTHROPIC_API_KEY"),
+}
+
+MODEL_MAP = {
+    "GPT-3.5": "gpt-3.5-turbo",
+    "GPT-4o": "gpt-4o",
+}
